Separate buffer packing logic
